requirements.txt
tkinterdnd2
PILLOW
requests
pyinstaller
pandas
openpyxl